fix: cast ttsDataArrayRef to Uint8Array for getByteFrequencyData method

+++ frontend/src/app/page.tsx
@@ -149,7 +149,7 @@ export default function Home() {
 
   const updateLipSync = () => {
     if (ttsAnalyserRef.current && ttsDataArrayRef.current) {
-      ttsAnalyserRef.current.getByteFrequencyData(ttsDataArrayRef.current);
+      ttsAnalyserRef.current.getByteFrequencyData(ttsDataArrayRef.current as Uint8Array);
       let sum = 0;
       for (let i = 0; i < ttsDataArrayRef.current.length; i++) {
         sum += ttsDataArrayRef.current[i];
